Updates to JobHost CORS support
