+++ src/WebJobs.Script.WebHost/Middleware/CorsMiddlewareFactory.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using Microsoft.AspNetCore.Cors.Infrastructure;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Azure.WebJobs.Script.WebHost.Configuration;
+using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
 namespace Microsoft.Azure.WebJobs.Script.WebHost.Middleware
@@ -13,21 +14,19 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.Middleware
     public class CorsMiddlewareFactory : ICorsMiddlewareFactory
     {
         private readonly IOptions<CorsOptions> _corsOptions;
+        private readonly ILoggerFactory _loggerFactory;
 
-        public CorsMiddlewareFactory(IOptions<CorsOptions> corsOptions)
+        public CorsMiddlewareFactory(IOptions<CorsOptions> corsOptions, ILoggerFactory loggerFactory)
         {
             _corsOptions = corsOptions;
+            _loggerFactory = loggerFactory;
         }
 
         public CorsMiddleware CreateCorsMiddleware(RequestDelegate next, IOptions<HostCorsOptions> corsOptions)
         {
-            CorsMiddleware middleware = null;
-            if (corsOptions?.Value != null)
-            {
-                var corsService = new CorsService(_corsOptions);
-                var policy = _corsOptions.Value.GetPolicy(_corsOptions.Value.DefaultPolicyName);
-                middleware = new CorsMiddleware(next, corsService, policy);
-            }
+            CorsPolicy policy = _corsOptions.Value.GetPolicy(_corsOptions.Value.DefaultPolicyName);
+            var corsService = new CorsService(_corsOptions);
+            var middleware = new CorsMiddleware(next, corsService, policy, _loggerFactory);
 
             return middleware;
         }
